refactor(native): split chara_detail_scene_scraper into .h/.cpp
Third and largest step of the header/.cpp split, applied to the scene scraper
(previously 1287 lines, all implementation in the header). Same shape as the
recognizer split.

The header keeps what must stay there:
- The updateUntilReady/readyAfterUpdate function templates.
- The FrameDescriptor and ImageOffsetEstimatorConfig data structs, the
  ScrapingInterpreter abstract interface, the ReadyState enum, and the
  CharaDetailSceneScraper static constexpr thresholds.
- Declarations of every scraper_impl class and their members (needed because
  they appear as by-value members and unique_ptr targets).

Every method body moves to the new .cpp verbatim, plus the file-local closeEnough
helper (now in an anonymous namespace).

Both consuming CMake source lists are updated in lockstep (native/CMakeLists.txt
and windows/runner/CMakeLists.txt).

Verified: umacapture_cli builds, umacapture_tests still passes (12/22),
`flutter build windows --debug` links the runner, and a full-pipeline video smoke
run (sandbox/inheritance_only_1.mp4, temporarily using the horizontal crop path)
drove the scraper end to end -- onCharaDetailStarted, onFactorProbe with 15
factors, onPageReady/onScrollReady per tab, and onCharaDetailFinished success:true.
